feat: extract CourseList from CourseTab
